% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/areas.R
\name{areas}
\alias{areas}
\title{Areas dataframe}
\format{A data frame with 7 variables and 323 observations.}
\source{
\url{xxxx}
}
\description{
Areas dataframe
}
\examples{
\dontrun{
  }
}
\references{
xxxx xxxx xxxx
}

