% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csPiUtils.R
\name{consistency}
\alias{consistency}
\title{Check the consistency between the slots of a csPi object}
\usage{
consistency(obj)
}
\arguments{
\item{obj:}{a csPi object}
}
\value{
a table
}
\description{
Check the consistency between the slots of a csPi object
}
\author{
Laurent Dubroca
}

