% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/COSTpk.R
\name{COSTPk}
\alias{COSTPk}
\title{Return key fields for a given COST data structure.}
\usage{
COSTPk(costTable = "tr")
}
\arguments{
\item{costTable}{the COST table.}
}
\value{
list of key fields for the given COST table.
}
\description{
Return key fields for a given COST data structure.
}
\examples{
#
}
\author{
Norbert Billet - IRD
}

