% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csAggregate.R
\name{csAggregate}
\alias{csAggregate}
\title{Splits the fishPi CS object into subsets, computes summary statistics for each, and returns the result in a convenient form.}
\usage{
csAggregate(csObj, x, by, aggFun, ...)
}
\arguments{
\item{csObj}{to aggregate.}

\item{x}{field to aggregate.}

\item{by}{a list of grouping elements.}

\item{aggFun}{a function to compute the summary statistic.}

\item{...}{further arguments passed to or used by aggFun}
}
\value{
A data frame with columns corresponding to the grouping variables in by followed by aggregated columns from x.

}
\description{
Splits the fishPi CS object into subsets, computes summary statistics for each, and returns the result in a convenient form.
}
\examples{
\dontrun{
data(sole)
sole.cs.pi <- csDataTocsPi(sole.cs)
aggByFoCatEu5 <- csAggregate(csObj=sole.cs.pi, x=list(lenNum="lenNum"), by=list(foCatEu5="foCatEu5", lenCls="lenCls"), sum, na.rm=TRUE)

library(ggplot2)
ggplot(data=aggByFoCatEu5) + geom_bar(mapping=aes(x=lenCls, y=lenNum), stat="identity") + facet_wrap(~foCatEu5, ncol=1, scales="free_y") + theme_bw()
}
}
\author{
Norbert Billet - IRD
}

