% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csformat.R
\name{csformat.xlsx}
\alias{csformat.xlsx}
\title{CS format definition}
\format{xlsx file}
\source{
WKRDB sharepoint \url{http://www.ices.dk}
}
\description{
An excel file defining the csPi format
included in the data directory of the fishPifct package
}

