% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapicesrect.R
\name{csMap}
\alias{csMap}
\title{This function map a variable from a slot from a csPi object}
\usage{
csMap(obj, slot, var, type = "tile", fct = "sum")
}
\arguments{
\item{obj:}{a csPi object}

\item{slot:}{the slot name}

\item{var:}{the variable name}

\item{type:}{the map type - tile or bubble. For categorical variable, only bubble is available.}

\item{fct:}{agregation function -sum, mean or n_distinct- to summarise the spatial information}
}
\value{
a ggplot plot
}
\description{
This function map a variable from a slot from a csPi object
}
\examples{
\dontrun{
library(fishPifct)
data(sole)
sole<- csDataTocsPi(sole.cs)
csMap(sole,"hh","foDur","tile","sum")
  	} 

  
}
\author{
Laurent Dubroca
}
\keyword{ICES}
\keyword{rectangle}
\keyword{statistical}

