% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/csSubset.R
\name{csSubset}
\alias{csSubset}
\title{Subset a fishPi CS object.}
\usage{
csSubset(csObj, subset, link = FALSE)
}
\arguments{
\item{csObj}{fishPi CS object to subset.}

\item{subset}{expression.}

\item{link:}{boolean TRUE if link to ca table is taken into account (merge of ca use only tr table)}
}
\value{
subseted fishPi CS object.
}
\description{
Subset a fishPi CS object.
}
\examples{
\dontrun{
data(sole)
library(ggplot2)
sole.cs.pi <- csDataTocsPi(sole.cs)
aggByFoCatEu5 <- csAggregate(csObj=sole.cs.pi, x=list(lenNum="lenNum"), by=list(foCatEu5="foCatEu5", lenCls="lenCls"), sum, na.rm=TRUE)
ggplot(data=aggByFoCatEu5) + geom_bar(mapping=aes(x=lenCls, y=lenNum), stat="identity") + facet_wrap(~foCatEu5, ncol=1, scales="free_y") + theme_bw()
}
}
\author{
Laurent Dubroca & Norbert Billet
}

