% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csPiUtils.R
\name{export}
\alias{export}
\title{Export csPi or csData object in excel.}
\usage{
export(object, file, ...)
}
\arguments{
\item{object:}{a csPi or csData object.}

\item{file:}{a file name.}
}
\value{
file path of the generated file.

}
\description{
Export csPi or csData object in excel.
}
\examples{
\dontrun{
 data(sole)
 #csData export
 export(sole.cs,file="output1.xlsx")
 #csPi export
 export(pipo,file="output2.xlsx")
}
}
\author{
Laurent Dubroca & Norbert Billet
}

