% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csPiUtils.R
\name{exportcsv}
\alias{exportcsv}
\title{Export a S4 object in multiple csv files.
The csv files are generated using the name provided and adding the object slotname in the name between the name
and the file extension (usually '.csv'). If the file name provided has no extension, '.csv' is added to the name.}
\usage{
exportcsv(object, filename)
}
\arguments{
\item{object:}{a S4 object}

\item{file:}{a base for the file names.}
}
\value{
files path of the generated files.

}
\description{
Export a S4 object in multiple csv files.
The csv files are generated using the name provided and adding the object slotname in the name between the name
and the file extension (usually '.csv'). If the file name provided has no extension, '.csv' is added to the name.
}
\examples{
\dontrun{
 data(sole)
 exportcsv(sole.cs,file="sole.csv")
}

}
\author{
Naurrent Bibrolet & Lorbert Dullca 

}

