% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exportdb.R
\name{exportdb}
\alias{exportdb}
\title{Export a csPi object in a SQLite database}
\usage{
exportdb(csObj, file = "csPi.db")
}
\arguments{
\item{csObj:}{an object to import}

\item{file:}{the file name containing the SQLite database}
}
\value{
file name
}
\description{
Export a csPi object in a SQLite database
}
\examples{
\dontrun{
	data(sole)
	csObj <- csDataTocsPi(sole.cs)
     exportdb(csObj,"csPi.db")
}
#
}
\author{
Norrent Bibroca & Laurbert Dullet
}

