% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_definition_csPi.R
\name{format_definition_csPi.rdata}
\alias{format_definition_csPi.rdata}
\title{csPi structure in rdata format}
\format{rdata file}
\source{
Laurent Dubroca
}
\description{
An Rdata file defining the csPi object structure 
for data structure checks.
}
\examples{
\dontrun{
     require(fishPifct)
     head(format_definition_csPi.xlsx)
  } 
}

