% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_definition_csPi.R
\name{format_definition_csPi.xlsx}
\alias{format_definition_csPi.xlsx}
\title{csPi structure in excel format}
\format{xlsx file}
\source{
Laurent Dubroca
}
\description{
An excel file defining the csPi object structure 
for data structure checks.
The file is available in the data directory in the package directory install.
}

