% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NB_datasets.R
\docType{data}
\name{format_RCM_MedBsLp}
\alias{format_RCM_MedBsLp}
\title{Format definition for RCM Med & BS & LP}
\format{A format definition.}
\usage{
data(format_RCM_MedBsLp)
}
\description{
The format used by RCM Med & BS & LP
}
\keyword{datasets}

