% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COSTpk.R
\name{fpKey}
\alias{fpKey}
\title{Return a string key composed by specified columns of a given table.}
\usage{
fpKey(tab, colIndex, sep = ":-:")
}
\arguments{
\item{tab}{the data table.}

\item{colIndex}{the key columns.}

\item{sep}{the characters pattern to separate fileds of the key.}
}
\value{
key.
}
\description{
Return a string key composed by specified columns of a given table.
}
\examples{
#
}
\author{
Norbert Billet - IRD
}

