% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csAggregate.R
\name{getCsJoinTable}
\alias{getCsJoinTable}
\title{Find the fishPi or CS tables to join to get all specified fields.}
\usage{
getCsJoinTable(csObj, fields)
}
\arguments{
\item{csObj}{fishPi or CS object.}

\item{fields}{vector of names of required fields.}
}
\value{
Vector of CS tables names to join.

}
\description{
Find the fishPi or CS tables to join to get all specified fields.
}
\examples{
getCsJoinTable(csObject, c("lenCls", "spp", "trpCode"))
getCsJoinTable(csObject, c("vslLen", "spp"))

}
\author{
Norbert Billet - IRD

}

