% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ICESAreaRects.R
\name{ICESAreaRects}
\alias{ICESAreaRects}
\title{ICES statistical rectangles and division}
\format{A data frame with 10 variables and 9167 observations.}
\source{
\url{http://www.ices.dk}
}
\description{
Data borrowed from COST packages
}
\examples{
\dontrun{
  }
}
\references{
xxxx xxxx xxxx
}

