% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/csPiUtils.R
\name{import}
\alias{import}
\title{Import csPi data from an excel file}
\usage{
import(file)
}
\arguments{
\item{file:}{an excel file name containing the csPi table.}
}
\value{
a csPi object
}
\description{
Import csPi data from an excel file
}
\examples{
\dontrun{
 data(sole)
 pipo<-csDataTocsPi(sole.cs)
 export(pipo,file="output.xslx")
}
}
\author{
Laurent Dubroca & Norbert Billet
}

