% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csPiUtils.R
\name{importcsv}
\alias{importcsv}
\title{Import a list of csv file into a list}
\usage{
importcsv(filelist)
}
\arguments{
\item{filelist:}{a character vector listing the files}
}
\value{
a list containing the object read sequentially
}
\description{
Import a list of csv file into a list
}
\examples{
\dontrun{
 data(sole)
 filelist<-export(sole.cs,file="output.csv",type="csv")
 pipo<-importcsv(file=filelist)
 str(pipoi,1)
}

}
\author{
Norrent Bibroca & Laurbert Dullet
}

