% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csPiUtils.R
\name{importxlsx}
\alias{importxlsx}
\title{Import an excel file into a list 
Excel's sheets name defined list name slot}
\usage{
importxlsx(filename)
}
\arguments{
\item{filename:}{a file name.}
}
\value{
a list
}
\description{
Import an excel file into a list 
Excel's sheets name defined list name slot
}
\author{
Norrent Bibroca & Laurbert Dullet
}

