% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mapicesrect.R
\name{mapicesrect}
\alias{mapicesrect}
\title{This function map a variable defined on ICES statistical rectangle}
\usage{
mapicesrect(data, var, rect)
}
\arguments{
\item{data:}{a data frame}

\item{var:}{the variable position (a numerical vector) or name (a character vector) in data. This variable should be numeric.}

\item{rect:}{the ICES statistical rectangle position (a numerical vector) or name (a character vector) in data}
}
\value{
a ggplot plot
}
\description{
This function map a variable defined on ICES statistical rectangle
}
\examples{
\dontrun{
	require(mapdata);require(dplyr);require(ggplot2)
	load("ICESAreaRects.rdata")
	datatmp<-data.frame(StatRect=sample(gsub(" ","",ICESAreaRects$StatRect),1000,replace=T),landWt=rnorm(1000,mean=1000,sd=300),month=sample(1:12,1000,replace=T))
	mapicesrect(datatmp,2,1)
  	}
}
\keyword{ICES}
\keyword{rectangle}
\keyword{statistical}

