% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/metier2vslType.R
\name{metier2vslType}
\alias{metier2vslType}
\title{This function compute vessel type according to metier and time spent fishing.}
\usage{
metier2vslType(fishpidata)
}
\arguments{
\item{fishpidata}{fishPi dataset}
}
\value{
Vector of the computed vessel type
}
\description{
Predominant fishing gear as used in the DCF fleet segmentation:
http://datacollection.jrc.ec.europa.eu/wordef/fleet-segment-dcf.
This should be unique for each vessel.
The dominance criteria shall be used to allocate each vessel to a segment based on the number of fishing days used
with each gear. If a fishing gear is used by more than the sum of all the others (i.e. a vessel spends more than 50 % of
its fishing time using that gear), the vessel shall be allocated to that segment. If not, the vessel shall be allocated to the
following fleet segment:
(a) ‘vessels using Polyvalent active gears’ if it only uses active gears;
(b) ‘vessels using Polyvalent passive gears’ if it only uses passive gears;
(c) ‘vessels using active and passive gears’.
}
\examples{
\dontrun{
  	}
}
\keyword{fishPi}

