% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csPiUtils.R
\name{outliers}
\alias{outliers}
\title{Outliers detection on a variable using
the adjusted outlyingness index (based on the robustbase package) if
the variable is numeric, and on a frequency table if not. 
In the later case, a treshold in percentage
is used to flag an outlier (1% by default: if a value is present less than the
treshold, then it is reported as outlier).}
\usage{
outliers(obj, slot, var, treshold = 1, ...)
}
\arguments{
\item{obj:}{a csPi object}

\item{slot:}{the slot where the variable of interest is located}

\item{var:}{the name of a variable on which outiers detection is applied}

\item{treshold:}{the percentage treshold to consider outliers in the case of non numeric data}

\item{other}{parameters related to the function adjOutlyingness from
the package robustbase}
}
\value{
plot a graphics and return a table of the table individuals flagged as outliers
}
\description{
Outliers detection on a variable using
the adjusted outlyingness index (based on the robustbase package) if
the variable is numeric, and on a frequency table if not. 
In the later case, a treshold in percentage
is used to flag an outlier (1% by default: if a value is present less than the
treshold, then it is reported as outlier).
}
\examples{
\dontrun{
 data(sole)
 sole<-csDataTocsPi(sole.cs)
 outliers(sole,"foDur")
}

}
\author{
Laurent Dubroca
}

