% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NB_codeList.R
\name{readCodeListFromXls}
\alias{readCodeListFromXls}
\title{Read code lists definitions from a xlsx file.}
\usage{
readCodeListFromXls(inputFilePath, formatNomenclatureName = FALSE,
  upperCaseCodes = TRUE)
}
\arguments{
\item{inputFilePath}{character: Path of the xlsx format definition file.}

\item{formatNomenclatureName}{logical: Should the names of the nomenclature to be formated ? (default: FALSE)}

\item{upperCaseCodes}{logical: Should the codes of the nomenclature to be formated in upper case ? (default: TRUE)}
}
\value{
List of nomenclatures.

}
\description{
Read code lists definitions from a xlsx file.
}
\examples{
\dontrun{
}
}
\author{
Norbert Billet
}

