% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NB_validation.R
\name{readFormatDbFromXls}
\alias{readFormatDbFromXls}
\title{Read a format definition from a xlsx def file.}
\usage{
readFormatDbFromXls(inputFilePath)
}
\arguments{
\item{inputFilePath}{character: Path of the xlsx format definition file.}
}
\value{
Format definition structure.

}
\description{
Read a format definition from a xlsx def file.
}
\examples{
\dontrun{
}
}
\author{
Norbert Billet
}

