% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NB_renderReports.R
\name{renderValidationReport}
\alias{renderValidationReport}
\title{Build a PDF or a HTML report on a data validation against a format definition.}
\usage{
renderValidationReport(obj, formatDb, reportFormat = "html", reportFilePath,
  title = "Validation of a dataset against a format", author = "John Doe",
  ...)
}
\arguments{
\item{obj}{Object to validate, see \code{\link{validateData}}.}

\item{formatDb}{Format data structure, see \code{\link{validateData}}.}

\item{reportFormat}{Format of the report, could be "html" (default) or "pdf".}

\item{reportFilePath}{File path of the output PDF or HTML document.}

\item{title}{Title of the report.}

\item{author}{Author of the report.}
}
\value{
File path of the created report.
}
\description{
Build a PDF or a HTML report on a data validation against a format definition.
}
\examples{

#
}
\author{
Norbert Billet - IRD
}

