% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sole.R
\name{sole.cs}
\alias{sole.cs}
\title{Sole commercial sampling data}
\format{cs, cl and ce data}
\source{
COSTcore package
}
\description{
A cs format file containing sole data for testing
coming from the COSTcore package
}

