% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NB_validation.R
\name{validateBio}
\alias{validateBio}
\title{Validate data against a format definition.}
\usage{
validateBio(obj)
}
\arguments{
\item{obj}{Object to validate, could be a data.frame or a S3/S4 object.}

\item{formatDb}{Format data structure.}

\item{ignoreCaseInCodelist}{logical: Should the case of codes of nomenclatures to be ignored ? (default: TRUE)}

\item{report}{character: Should the report be saved in .csv file (\"files\"), returned in a R session list (\"list\") or both (\"both\") ? (default: \"files\")}

\item{reportDir}{character: Path where to record the report files, otherwise the path of the per-session temporary directory is used.}
}
\value{
Validation report.
}
\description{
Validate data against a format definition.
}
\examples{
\dontrun{
}
}
\author{
Norbert Billet
}

