% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csSubset.R
\name{csSubset}
\alias{csSubset}
\title{Subset a fishPi CS object.}
\usage{
csSubset(csObj, subset)
}
\arguments{
\item{csObj}{fishPi CS object to subset.}

\item{subset}{expression.}
}
\value{
subseted fishPi CS object.

}
\description{
Subset a fishPi CS object.
}
\examples{
\dontrun{
data(sole)
piSole.cs <- csDataTocsPi(sole.cs)
piSole.cs.sub <- csSubset(piSole.cs, daysAtSea > 5)

}
}
\author{
Laurent Dubroca & Norbert Billet

}

