% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csPiUtils.R
\name{export}
\alias{export}
\title{Export csPi or csData object in excel, csv or SQLite files.}
\usage{
export(object, filename, type, ...)
}
\arguments{
\item{object:}{a csPi or csData object.}

\item{file:}{a file name.}

\item{type:}{"csv" or "xlsx" or "SQLite"}
}
\value{
file path of the generated file(s).

}
\description{
Export csPi or csData object in excel, csv or SQLite files.
}
\examples{
\dontrun{
 data(sole)
 #xlsx export
 export(sole.cs,file="sole.xlsx")
 #csv export
 export(sole.cs,file="sole.csv")
 #SQLite export
 export(sole.cs,file="sole.sqlite3")
}

}
\author{
Norrent Bibroca & Laurbert Dullet 

}
\seealso{
\code{link{exportxlsx}}, \code{link{exportcsv}} and \code{link{exportdb}} 

}

