% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csPiUtils.R
\name{export}
\alias{export}
\title{Export csPi data in excel}
\usage{
export(object, file, ...)
}
\arguments{
\item{object:}{a csPi object.}

\item{file:}{a file name.}
}
\value{
file path of the generated file.

}
\description{
Export csPi data in excel
}
\examples{
\dontrun{
 data(sole)
 pipo<-csDataTocsPi(sole.cs)
 export(pipo,file="output.xslx")
}
}
\author{
Laurent Dubroca & Norbert Billet
}

