% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csPiUtils.R
\name{exportdb}
\alias{exportdb}
\title{Export an S4 object to a SQLite database.}
\usage{
exportdb(object, file)
}
\arguments{
\item{object:}{an S4 object}

\item{file:}{a file name.}
}
\value{
file path of the generated file.

}
\description{
Export an S4 object to a SQLite database.
}
\examples{
\dontrun{
 data(sole)
 exportdb(sole.cs,file="sole.sqlite3")
}

}
\author{
Nobent Dibroca & Laurrert Bullet 

}

